+++ CHANGELOG.md
@@ -16,6 +16,8 @@ This major release introduces a few breaking changes. We've provided more inform
 - Added support for context parallelism.
 - Added support for expert parallelism with MoE models.
 - Added in-loop evals for Minerva, GSM, HumanEval, MBPP (`ai2-olmo-eval==0.7.0`)
+- Added `CosWithWarmupAndLinearDecay` learning rate scheduler
+- Added `WSD` learning rate scheduler
 
 ### Changed
 
@@ -27,6 +29,10 @@ This major release introduces a few breaking changes. We've provided more inform
 - Removed the following callbacks: `MoEHandlerCallback`, `SchedulerCallback`, `MatrixNormalizerCallback`, `GradClipperCallback`, and `Float8HandlerCallback`.
   The functionality from all of those callbacks has been moved to the `TransformerTrainModule` class.
 - Removed the callback methods `.pre_eval_batch()` and `.post_eval_batch()`.
+
+### Fixed
+
+- Fixed the model ladder code when training on mps or cpu device
 
 ## [v1.9.0](https://github.com/allenai/OLMo-core/releases/tag/v1.9.0) - 2025-03-10